requirements.txt
Django==2.1.5
pytz==2018.9
tqdm==4.29.1
requests==2.21.0
django-filter==2.0.0
psycopg2==2.7.6.1